Make ParameterItem.equals(Any?)/hashCode() consistent across models
Previously, `TextParameterItem` implemented equality solely in terms of
its `parameterIndex`, whereas `TurbineParameterItem` and
`PsiParameterItem` implemented it in terms of `parameterIndex` and
`containingMethod`. However, investigation showed that the method was
only used in `ComparatorVisitor` to compare parameters within the same
method so checking `containingMethod` is redundant.

This change follows a similar pattern to that used in `MethodItem` to
dedup them and make them consistent in only comparing `parameterIndex`.

Bug: 351604449
Test: ./gradlew
Change-Id: Idc01e150e83b7942fa0f41875d1b7d7d888f91cc