Make createDefaultConstructor consistent in text and turbine
The text implementation is not actually called so where there is a
functional difference, e.g. visibility, defer to turbine. Although the
text implementation is not used it is kept for now to minimize diffs.
Once it has been merged with the turbine version into a
`DefaultConstructorItem` it will be called just as the turbine one
currently is.

Bug: 351410134
Test: ./gradlew
Change-Id: Ie320f0fc73effec5ce0725b46c49c92cb6c6d7d4