Add ParameterItem.duplicate(MethodItem, TypeParameterBindings)
Previously, the method was defined on `TextParameterItem` and
`TurbineParameterItem` separately which meant any code that called it
either had to use the model specific type or cast it to the model
specific type. This change adds it to `ParameterItem` allowing the
model specific types to be replaced with `ParameterItem` making the
follow-up change easier.

At the moment `PsiParameterItem` does not provide its own
implementation as it duplicates `MethodItem` by reconstructing from
the `PsiMethod/PsiParameter` objects. That may change in future but is
not needed right now.

Bug: 351410134
Test: ./gradlew
Change-Id: I2597a97ef9b89473e3eae22ff667d7858a33e98e