Make MethodItem.duplicate(ClassItem) consistent across all models
Previously, `TextMethodItem` would map type variables in the return and
parameter type to match the `ClassItem` into which the method was being
duplicated but the Psi and Turbine models did not. This change makes
them behave in the same way as the text model. That behavior is better
because it is more accurate.

It was necessary as part of this to not map type variables in
`PsiMethodItem.duplicate(ClassItem)` if it is being used to duplicate a
method into a non-PsiBasedCodebase to avoid possible class cast
exceptions in `PsiParameterItem.duplicate(...)` when it casts the
converted `TypeItem` into a `PsiTypeItem`.

Bug: 351410134
Test: ./gradlew
Change-Id: Ie76a1b7d2dc3455f792ee33318a10ef1ddd1fea9