Remove ClassItem.inheritMethodFromNonApiAncestor(MethodItem)
Previously, it was just a wrapper that switched the parameter and
receiver on `MethodItem.duplicate(ClassItem)`. It added no value to
the model so, this change inlines it and removes it. A number of tests
also needed updating and they were either converted to test `duplicate`
or removed if the tests already `duplicate`.

It was only used from a single call site in `ApiAnalyzer` and the same
`Codebase` test was not necessary there as it is only called with a
single `Codebase` so it was removed while inlining.

Bug: 351410134
Test: ./gradlew
Change-Id: I1a8ac3978a5e24182cdbaadd21fff26bef6bcce0