Implement allInterfaces() consistently across the models
Each model has their own, slightly different implementation of the
`ClassItem.allInterfaces()` method. The text one is incorrect, the Psi
and Turbine appear to be identical in behavior but it is difficult to
determine as the Psi one uses a mixture of IntelliJ Psi and Psi model
functionality. This change makes them all use the Turbine one as it is
simpler and better explained.

Bug: 351410134
Test: ./gradlew
Change-Id: I4fefea93443fdd1920579333972a306369dc5ced