Add CallableItem as a super interface of MethodItem
Moves methods from `MethodItem` that clearly also apply to
`ConstructorItem`. It is possible that other methods may need moving in
future. This purposely kept the same order so as to make it easier to
diff against the original `MethodItem` and does not change any "method"
references. They will be done in a follow up change where it will be
easier to see the changes.

Bug: 352481366
Test: ./gradlew
Change-Id: Ia322f7ad48114cec497b33403c8c7acc4162ec21