Remove PackageItem.isDefault property as it is not needed
Previously, it was only used in `ApiAnalyzer` to avoid inheriting the
hidden status of the containing package if it was the root package
however, that did nothing because the root package is never marked as
hidden.

For reference https://r.android.com/757573 was the change that added
the `isDefault` property to `PackageItem` and used it in `ApiAnalyzer`.

Bug: 351118278
Test: ./gradlew
Change-Id: I80ed2403d109e98a89eecff7611071fc3258a7ba