Stop disabling filtering of super type on TextClassItem
Historically, since the first Metalava drop, `TextClassItem` has
overridden `filteredSuperClassType(Predicate)` to prevent it from
doing any filtering. That only appears to have an impact on the
`signature-to-jdiff` sub-command that reads a signature file and then
strips it so it only refers to types defined within the signature file.
However, due to the override it did not filter out the super class type
which meant that the `extends` attribute would refer to a type that was
not defined.

Unfortunately, JDiff files are created and processed for a number of
different purposes so it is extremely difficult to determine if
filtering the super class type would break the consumers. So, this
change removes the override, which is needed to allow the
`TextClassItem` to be replaced with `DefaultClassItem`, and then adds
an option to `FilteringApiVisitor` to disable filtering of the super
type. That option is used in `SignatureToJDiffCommand` to preserve the
legacy behavior.

Bug: 351410134
Test: ./gradlew
      # Also tested to make sure that removing the override changed the
      # output of some files in the Android build and disabling the
      # filtering of the super type in `SignatureToJDiffCommand`
      # stopped the change.
Change-Id: I1410ab3c87a9840472793d4b9361b3a2131666e8