Add PsiCallableItem
Previously, `DefaulParameterItem` was switched to callable before the
`DefaultCallableItem` was created but this is creating
`PsiCallableItem` before switching `PsiParameterItem`. That is because
the casting in the creation of `PsiParameterItem`s means it does not
need to be done first and doing this first simplifies the changes
because it allows the `ParameterItemsFactory` to be changed to
`PsiCallableItem` at the same time as switching `PsiParameterItem`
instead of after.

Bug: 352481366
Test: ./gradlew
Change-Id: Ifd11260ebf4b3729ad3a9f40ff38ff3aafac8668