Maintain a list of error Report objects not formatted messages
`DefaultReporter` maintains a list of errors so that it can reprint
some at the end to make them easier to find. This change switches that
from a list of `String` to a list of `Report` so that they can be
sorted in the same way as other `Report` objects will be sorted.

Bug: 352868158
Test: ./gradlew
Change-Id: I9d1e742a960cd637abfdc7eb7c47d4d9169d6f4a