Add InheritableItem and implement by FieldItem and MethodItem
Previously, every `MemberItem` subclass implemented `inheritedFrom`
and `duplicate()` even if it could not actually be inherited, i.e.
property and constructor. This creates a separate `InheritableItem`
interface that is only implemented by `FieldItem` and `MethodItem`.
While that still means `ConstructorItem` implements it as currently it
extends `MethodItem` in future that will not be the case at which point
`ConstructorItem` will not be an `InheritableItem` either.

Bug: 352481366
Test: ./gradlew
Change-Id: I1941e3cda9fb9e8b2ee26390d8a1f869f834941d