Remove unnecessary uses and implementations of isConstructor
The `isConstructor()` method only makes sense when called on
`CallableItem` as it is always `true` when called on `ConstructorItem`
and `false` when called on `MethodItem`. This change moves all the
implementations to the interfaces and deprecates the `ConstructorItem`
and `MethodItem` overrides so that the compiler reports uses of them.
That made it easy to remove the unnecessary ones and will prevent any
additional ones being added in future.

Bug: 352481366
Test: ./gradlew
Change-Id: I151f3e9fbb1361238bdeff9ce91b0b273b54a974