Stop filtering no-args constructor
The Kotlin compiler generates a no-argument constructor when the primary constructor parameters all have default values, even when @JvmOverloads isn't used.

The no-arg constructor was being filtered out during psi codebase creation because it is implicit, but since it is only generated based on the primary constructor, it could previously be silently removed if the primary constructor no longer had all default values, which would break compatibility, so it should be tracked to make sure that doesn't happen.

This requires temporarily disabling androidx_with_metalava to add the no-arg constructors to signature files (aosp/3180142), but since this change only impacts kotlin-source it should not impact platform.

Bug: 352776539
Test: ApiFileTest, CompatibilityCheckTest
Change-Id: I223af60dd90d15e0dcb6457f0f45fcee55027385