CompatibilityCheck.compare(MethodItem...) group callable checks
Previously, the `compare(MethodItem, MethodItem)` method had a mixture
of checks some of which applied to callables, i.e. method and
constructors, but most of which applied only to methods. The latter did
produce some incorrect issues, like reporting a change in the `final`
modifier of a constructor even though `final` is not supported on
constructors.

This change groups the callable checks at the start and them exits
immediately after if the `MethodItem` is actually a `ConstructorItem`.
That broke a few tests that incorrectly reported `AddedFinal` on
constructors so this change removes those issues that are no longer
reported.

One test relied on the `RemovedFinalStrict` being reported on a
constructor so that was modified to use a method instead.

Bug: 352481366
Test: ./gradlew
Change-Id: I609cb8d5c59aeaa88eea804d5f547f63b8d2ae88