Use inheritableHidden to support --show-single-annotation
A `--show-single-annotation` annotation (non-recursive show) is
supposed to unhide the annotated `Item` while not affecting the hidden
status of its contents. Previously, the way that was supported was the
`Item.hidden` was set as for a recursive show annotation but the
children, e.g. nested classes and members had to check the parent's
`originallyHidden` and `showability` to see if it was hidden because of
a non-recursive show annotation.

That was complicated, required duplicating code for each child `Item`
type (since deduped) and also required an additional check in the
inheritance logic.

This change uses the `inheritableHidden` property added previously to
pull the differentiation between a recursive and non-recursive show
into the `Item` to which it is attached.

Setting each of the types of show annotation on a hidden `Item` has
the following effect:
* Recursive:
  * `originallyHidden=true`
  * `inheritableHidden=false`
  * `hidden=false`

* Non-recursive:
  * `originallyHidden=true`
  * `inheritableHidden=true`
  * `hidden=false`

Bug: 351118278
Test: ./gradlew
Change-Id: I10670078127ad80953487e92aff48b9a9ca8d46d