Stop unnecessary package traversal in ApiPredicate
Previously, the `ApiPredicate` needed to traverse up the package
hierarchy checking for hidden packages to make sure that it did not
mistakenly include classes that were in a hidden package but loaded
after `propagateHiddenRemovedAndDocOnly()` ran in the API.

A specific case is the `com.android.internal.util.DataClass` annotation
class. It is in a hidden package and is not provided in the source
files passed to Metalava on the command line, so it is not loaded
before `propagateHiddenRemovedAndDocOnly()` is called. However, it is
provided in the source files available for Psi to find so is not on the
class path.

The class is loaded when `AnnotationItem.resolve()` is called. Because
that was after `propagateHiddenRemovedAndDocOnly()` it was not marked
as hidden even though it was in a hidden package. So, the
`ApiPredicate`'s package traversal compensated for that.

However, the inheritance is done automatically now when requested so
the `DataClass`'s `hidden` property is correctly set to `true` because
it is in a hidden package and so the `ApiPredicate` no longer needs to
compensate for that so this change removes it.

Bug: 351118278
Test: ./gradlew
Change-Id: Ie5c4729e4cdd23d99f61643a2f225a147e39d722