Add SelectableItem marker interface
Currently, every `Item` supports properties like `hidden` and `removed`
which determine whether the `Item` will be selected to be part of an
API. However, not every `Item` can be individually selected, i.e.
`ParameterItem` and `TypeParameterItem` will be selected if and only if
their containing `Item` is selected. That means it makes no sense to
have them support those properties and doing so complicates the code
that does that selection and makes it harder to reason about.

This change is the start of a series of changes that will refactor the
code to correctly differentiate between the selectable and
non-selectable `Item`s.

Bug: 351118278
Test: ./gradlew
Change-Id: I096c5f4ad454c03e702b7743cba8d5b23fd376bf