Refactor tracking of ApiVariantSelectors lazy boolean properties
Previously, it encoded two bits for each property (one to indicate
whether it has been set and one for the value), side-by-side within
`flags`. That meant defining and passing around two bit masks for each.
This change switches to store those bits at the same position in two
separate flags. That simplifies the use of the `lazyGet()` and
`lazySet()` functions and also made it easier to add a `toString()
method.

The main purpose of using bit masks is not performance or storage but
rather to support lazy initialization with optional setters without
duplicating lots of complicated code.

Bug: 351118278
Test: ./gradlew
Change-Id: Ia8ab958233c38ae93e3d0b15c53dfed3582736f3