Move ensureParentIsVisibleIfThisIsVisible() into ApiVariantSelectors
Previously, it was separated out from the main inheritance logic in
https://r.android.com/3180681. That was a mistake as the inheritance
logic needs to fix up the `hidden` properties afterwards so that it
matches the `ApiPredicate` behavior.

This change effectively reverts the previous change albeit, with the
`ensureParentVisible()` method being in `ApiVariantSelectors.Mutable`
instead of `ApiAnalyzer`.

The `SHOWING_MEMBER_IN_HIDDEN_CLASS` issue needed to be moved from the
default category, i.e. `Category.UNKNOWN` to `Category.API_LINT`. That
is because previously it was reported from the `ApiAnalyzer` which is
passed the `reporterApiLint`. However, this change causes it to be
reported via the Codebase.reporter` which is a
`CategoryRedirectingReporter` that selects the appropriate `Reporter`
to use based on the issue's category. For reference that was added in
https://r.android.com/3182981.

Bug: 351118278
Test: ./gradlew
Change-Id: I3246fb12077d389e85450e8670ba4a52b23881c9